cast numerator to double

+++ 02-basic-calculator-api/02-basic-calculator-api/src/main/java/com/codewithudo/basiccalculatorapi/controllers/CalculatorController.java
@@ -42,7 +42,7 @@ public class CalculatorController {
                     "error", "Cannot divide by zero"
             );
         }
-        double dividend = a / b;
+        double dividend = (double) a / b;
         return Map.of(
                 "operation", "division",
                 "result", dividend);
